+++ CHANGELOG.md
@@ -8,6 +8,9 @@ AppEngine version, listed here to ease deployment and troubleshooting.
    * new entities do not have the user in their key
    * WARNING: user merge on pending old consent entities does not work
  * Removed support for old uploader invites.
+ * Added `backfill_packageisflutterfavorite.dart` to be executed before and
+   after next deployment, in order to backfill the `Package.isFlutterFavorite`
+   property.
 
 ## `20191104t103859-all`
  * Refactored `Consent`: